+++ src/main/java/mx/nic/lab/rpki/sqlite/object/SlurmBgpsecDbObject.java
@@ -40,7 +40,7 @@ public class SlurmBgpsecDbObject extends SlurmBgpsec implements DatabaseObject {
 		if (resultSet.wasNull()) {
 			setId(null);
 		}
-		setAsn(resultSet.getInt(ASN_COLUMN));
+		setAsn(resultSet.getLong(ASN_COLUMN));
 		if (resultSet.wasNull()) {
 			setAsn(null);
 		}
