+++ src/main/java/mx/nic/lab/rpki/sqlite/object/SlurmPrefixDbObject.java
@@ -43,7 +43,7 @@ public class SlurmPrefixDbObject extends SlurmPrefix implements DatabaseObject {
 		if (resultSet.wasNull()) {
 			setId(null);
 		}
-		setAsn(resultSet.getInt(ASN_COLUMN));
+		setAsn(resultSet.getLong(ASN_COLUMN));
 		if (resultSet.wasNull()) {
 			setAsn(null);
 		}
